Fix Broken Postgres Tests
